refactor server trace
